# api/wegia/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Wegia",
        "description": "Documentação da API Wegia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor"
        }
    ],
    "paths": {
...
        "/atendido": {
            "get": {
                "tags": [
                    "Atendido"
                ],
                "summary": "Buscar os Atendidos",
                "operationId": "3796b9d1a57bae33208566688597ca1a",
                "parameters": [
                    {
                        "name": "id_status",
                        "in": "query",
                        "description": "Id do status do atendido",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "buscar",
                        "in": "query",
                        "description": "Nome ou CPF do atendido para busca",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ordenacao",
                        "in": "query",
                        "description": "Campo para ordenar (nome e cpf)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipoOrdenacao",
                        "in": "query",
                        "description": "Tipo de ordenação",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "ASC"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Separados por virgula",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "itensPorPagina",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "pagina",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacao realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Atendido"
                ],
                "summary": "Cadastrar um novo Atendido",
                "operationId": "ca7f52a2de8c57458f463f71fb521175",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AtendidoCadastrarValidation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cadastro realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/atendido/{id}": {
            "get": {
                "tags": [
                    "Atendido"
                ],
                "summary": "Buscar Atendido por ID",
                "operationId": "b96e52ff6cbef9a6cc4e1cd2750fcd92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do  atendido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Separados por virgula",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacao realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Atendido"
                ],
                "summary": "Buscar Atendido por ID",
                "operationId": "32ebdac098ac748819c1a75f8b7b393c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do  atendido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AtendidoAtualizarValidation"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Operacao realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/configuracao/campo-imagem": {
            "get": {
                "tags": [
                    "Configuracao"
                ],
                "summary": "Buscar os campos de imagens",
                "operationId": "eda529cc6e098b2d11d1569acaab0ac7",
                "responses": {
                    "200": {
                        "description": "Operacao realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AtendidoAtualizarValidation": {
                "required": [
                    "atendido_status_idatendido_status"
                ],
                "properties": {
                    "atendido_status_idatendido_status": {
                        "description": "id do status do atendido",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "AtendidoCadastrarValidation": {
                "required": [
                    "pessoa_id_pessoa",
                    "atendido_tipo_idatendido_tipo",
                    "atendido_status_idatendido_status"
                ],
                "properties": {
                    "pessoa_id_pessoa": {
                        "description": "id da pessoa",
                        "type": "integer"
                    },
                    "atendido_tipo_idatendido_tipo": {
                        "description": "id do tipo do atendido",
                        "type": "integer"
                    },
                    "atendido_status_idatendido_status": {
                        "description": "id do status do atendido",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}